requirements.txt
python-magic==0.4.27
scikit-learn==1.3.0
nltk==3.8.1
tqdm==4.66.1
pyinstaller==6.3.0
python-docx==1.0.1
openpyxl==3.1.2
PyPDF2==3.0.1
Pillow==10.2.0
python-pptx==0.6.23 